+++ NestSuite/AutomationIds.cs
@@ -1,0 +1,106 @@
+namespace NestSuite;
+
+public static class AutomationIds
+{
+    public static class Shell
+    {
+        public const string TabStrip             = "Shell.TabStrip";
+        public const string TabAddButton         = "Shell.TabAddButton";
+        public const string TabListButton        = "Shell.TabListButton";
+        public const string NoteNestWorkspace    = "Shell.NoteNestWorkspace";
+        public const string ChatNestWorkspace    = "Shell.ChatNestWorkspace";
+        public const string IdeaNestWorkspace    = "Shell.IdeaNestWorkspace";
+        public const string TempNestWorkspace    = "Shell.TempNestWorkspace";
+    }
+
+    public static class NoteNest
+    {
+        public const string FilterBox                 = "NoteNest.FilterBox";
+        public const string FilterClearButton         = "NoteNest.FilterClearButton";
+        public const string NotebookTree              = "NoteNest.NotebookTree";
+        public const string EditorHost                = "NoteNest.EditorHost";
+        public const string RightGridSplitter         = "NoteNest.RightGridSplitter";
+        public const string RightPaneExpandButton     = "NoteNest.RightPaneExpandButton";
+        public const string RightPaneTabControl       = "NoteNest.RightPaneTabControl";
+        public const string MarkerSortComboBox        = "NoteNest.MarkerSortComboBox";
+        public const string TodoFilterCheckBox        = "NoteNest.TodoFilterCheckBox";
+        public const string FixmeFilterCheckBox       = "NoteNest.FixmeFilterCheckBox";
+        public const string NoteMarkerFilterCheckBox  = "NoteNest.NoteMarkerFilterCheckBox";
+        public const string AddNotebookButton         = "NoteNest.AddNotebookButton";
+        public const string AddNoteButton             = "NoteNest.AddNoteButton";
+        public const string InsertTodoButton          = "NoteNest.InsertTodoButton";
+        public const string InsertFixmeButton         = "NoteNest.InsertFixmeButton";
+        public const string InsertNoteButton          = "NoteNest.InsertNoteButton";
+    }
+
+    public static class Editor
+    {
+        public const string EditorBox      = "Editor.EditorBox";
+        public const string LineNumberBox  = "Editor.LineNumberBox";
+        public const string CompletionList = "Editor.CompletionList";
+    }
+
+    public static class IdeaNest
+    {
+        public const string SearchBox     = "IdeaNest.SearchBox";
+        public const string CardArea      = "IdeaNest.CardArea";
+        public const string AddIdeaButton = "IdeaNest.AddIdeaButton";
+    }
+
+    public static class ChatNest
+    {
+        public const string SearchBox            = "ChatNest.SearchBox";
+        public const string InputBox             = "ChatNest.InputBox";
+        public const string PostButton           = "ChatNest.PostButton";
+        public const string SpeakerSelf          = "ChatNest.SpeakerSelf";
+        public const string SpeakerRefute        = "ChatNest.SpeakerRefute";
+        public const string SpeakerSupplement    = "ChatNest.SpeakerSupplement";
+        public const string SpeakerConclusion    = "ChatNest.SpeakerConclusion";
+        public const string ScrollToBottomButton = "ChatNest.ScrollToBottomButton";
+    }
+
+    public static class TempNest
+    {
+        public const string Slot1TitleBox   = "TempNest.Slot1.TitleBox";
+        public const string Slot1BodyBox    = "TempNest.Slot1.BodyBox";
+        public const string Slot1CopyButton = "TempNest.Slot1.CopyButton";
+        public const string Slot1ClearButton = "TempNest.Slot1.ClearButton";
+        public const string Slot2TitleBox   = "TempNest.Slot2.TitleBox";
+        public const string Slot2BodyBox    = "TempNest.Slot2.BodyBox";
+        public const string Slot2CopyButton = "TempNest.Slot2.CopyButton";
+        public const string Slot2ClearButton = "TempNest.Slot2.ClearButton";
+        public const string Slot3TitleBox   = "TempNest.Slot3.TitleBox";
+        public const string Slot3BodyBox    = "TempNest.Slot3.BodyBox";
+        public const string Slot3CopyButton = "TempNest.Slot3.CopyButton";
+        public const string Slot3ClearButton = "TempNest.Slot3.ClearButton";
+        public const string Slot4TitleBox   = "TempNest.Slot4.TitleBox";
+        public const string Slot4BodyBox    = "TempNest.Slot4.BodyBox";
+        public const string Slot4CopyButton = "TempNest.Slot4.CopyButton";
+        public const string Slot4ClearButton = "TempNest.Slot4.ClearButton";
+    }
+
+    public static class Dialog
+    {
+        public const string FindBox            = "Dialog.FindBox";
+        public const string ReplaceBox         = "Dialog.ReplaceBox";
+        public const string FindNextButton     = "Dialog.FindNextButton";
+        public const string FindPrevButton     = "Dialog.FindPrevButton";
+        public const string ReplaceButton      = "Dialog.ReplaceButton";
+        public const string ReplaceAllButton   = "Dialog.ReplaceAllButton";
+        public const string CloseButton        = "Dialog.CloseButton";
+        public const string CaseCheckBox       = "Dialog.CaseCheckBox";
+        public const string AllNotesCheckBox   = "Dialog.AllNotesCheckBox";
+        public const string AllNotesResultList = "Dialog.AllNotesResultList";
+        public const string FontFamilyBox      = "Dialog.FontFamilyBox";
+        public const string FontSizeBox        = "Dialog.FontSizeBox";
+        public const string OkButton           = "Dialog.OkButton";
+        public const string CancelButton       = "Dialog.CancelButton";
+        public const string PromptLabel        = "Dialog.PromptLabel";
+        public const string InputBox           = "Dialog.InputBox";
+        public const string NoteFilterBox      = "Dialog.NoteFilterBox";
+        public const string NoteList           = "Dialog.NoteList";
+        public const string ResultList         = "Dialog.ResultList";
+        public const string NavigateButton     = "Dialog.NavigateButton";
+        public const string HeaderText         = "Dialog.HeaderText";
+    }
+}